Major refactor of player store
