+++ src/Pages/Players/ListOfPlayers/TableWithPlayersDialog.tsx
@@ -3,7 +3,7 @@ import { useDispatch } from 'react-redux';
 
 import DeleteDialog from '../../../Components/Communication/DeleteDialog';
 import { IPlayer } from '../../../models/player.model';
-import { playerDelete } from '../../../store/players/actions';
+import { playerDelete } from '../../../store/players/actions-creators';
 
 interface IProps {
   playerToDelete: IPlayer;
