+++ src/store/players/action-types.ts
@@ -1,0 +1,8 @@
+export enum ActionType {
+  ADD_NEW_PLAYER = 'add new player',
+  DELETE_PLAYER = 'delete player',
+  FETCH_ALL_PLAYERS = 'fetch all players',
+  FETCH_PLAYER_PROFILE = 'fetch player profile',
+  FETCH_PLAYER_SCORES = 'fetch player scores',
+  UPDATE_ADMIN_STATUS = 'update admin status',
+}
