+++ src/store/user/actions.ts
@@ -7,7 +7,7 @@ import { apiUrl } from '../../config/constants';
 import { ILogInCredentials, IProfileDetails } from '../../models/credentials.model';
 import { IUser } from '../../models/player.model';
 import { appDoneLoading, appLoading, setMessage } from '../appState/actions';
-import { removeAllPlayers } from '../players/actions';
+import { removeAllPlayers } from '../players/actions-creators';
 import { removeAllFixtures } from '../predictions/actions';
 import { removeAllScores } from '../scores/actions';
 import { removeAllTeams } from '../teams/actions';
